Prevent future warning in pd.concat
Warning:

The behavior of DataFrame concatenation with empty
or all-NA entries is deprecated. In a future version, this
will no longer exclude empty or all-NA columns when
determining the result dtypes. To retain the old behavior,
exclude the relevant entries before the concat operation.

Solution: Check if the data frame is empty before concat

Source: https://stackoverflow.com/a/77254951